refactor(runtime): keep late-result validation within package baselines
