+++ pkg/services/factory_runtime/internal/services/orchestration/subsystems/subsystem_transitioner_failure_resource_test.go
@@ -346,3 +346,276 @@ func TestCalculateMutations_AcceptedMixedWorkResource_ReleasesConsumedResourceRe
 		})
 	}
 }
+
+// portos:func-length-exception owner=agent-factory reason=legacy-resource-release-fixture review=2026-07-18 removal=split-resource-setup-and-failure-release-assertions-before-next-history-transitioner-change
+func TestHistoryTransitionerPipeline_FailureReleasesConsumedResourceTokenIdentity(t *testing.T) {
+	tp, snapshot, resourceConsumed := newFailureReleasesConsumedResourceFixture(false)
+	result, err := tp.Execute(context.Background(), snapshot)
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+	assertFailedMixedWorkResourceRelease(t, result, resourceConsumed)
+}
+
+func TestHistoryTransitionerPipeline_FailureReleasesConsumedResourceRegardlessOfInputOrder(t *testing.T) {
+	orderings := []struct {
+		name      string
+		workFirst bool
+	}{
+		{name: "resource-first", workFirst: false},
+		{name: "work-first", workFirst: true},
+	}
+
+	for _, ordering := range orderings {
+		t.Run(ordering.name, func(t *testing.T) {
+			tp, snapshot, resourceConsumed := newFailureReleasesConsumedResourceFixture(ordering.workFirst)
+			result, err := tp.Execute(context.Background(), snapshot)
+			if err != nil {
+				t.Fatalf("unexpected error: %v", err)
+			}
+			assertFailedMixedWorkResourceRelease(t, result, resourceConsumed)
+		})
+	}
+}
+
+func newFailureReleasesConsumedResourceFixture(workFirst bool) (*testPipeline, *interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net], factorytoken.Token) {
+	n := &state.Net{
+		Places: map[string]*petri.Place{
+			"wt-code:init":       {ID: "wt-code:init", TypeID: "wt-code", State: "init"},
+			"wt-code:failed":     {ID: "wt-code:failed", TypeID: "wt-code", State: "failed"},
+			"executor:available": {ID: "executor:available", TypeID: "executor", State: "available"},
+		},
+		Transitions: map[string]*petri.Transition{
+			"t1": {
+				ID:         "t1",
+				Name:       "code",
+				WorkerType: "agent",
+				InputArcs: []petri.Arc{
+					{ID: "a1", Name: "work", PlaceID: "wt-code:init", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
+					{ID: "a2", Name: "resource", PlaceID: "executor:available", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
+				},
+				FailureArcs: []petri.Arc{
+					{ID: "a3", Name: "fail", PlaceID: "wt-code:failed", Direction: petri.ArcOutput},
+				},
+			},
+		},
+		WorkTypes: map[string]*state.WorkType{
+			"wt-code": {
+				ID: "wt-code",
+				States: []state.StateDefinition{
+					{Value: "init", Category: state.StateCategoryInitial},
+					{Value: "failed", Category: state.StateCategoryFailed},
+				},
+			},
+		},
+	}
+	tp := newTestPipeline(n, nil)
+	tp.WriteResult(workerexecution.WorkResult{DispatchID: "d-1", TransitionID: "t1", Outcome: workerexecution.OutcomeFailed, Error: "agent crashed"})
+
+	now := time.Date(2026, time.April, 6, 11, 0, 0, 0, time.UTC)
+	resourceConsumed := factorytoken.Token{
+		ID:        "executor:resource:0",
+		PlaceID:   "executor:available",
+		CreatedAt: now.Add(-3 * time.Hour),
+		EnteredAt: now.Add(-3 * time.Hour),
+		Color: factorytoken.Color{
+			WorkID:     "executor:0",
+			WorkTypeID: "executor",
+			DataType:   factorytoken.DataTypeResource,
+			Tags:       map[string]string{"pool": "shared"},
+		},
+		History: factorytoken.History{
+			PlaceVisits: map[string]int{"executor:available": 4},
+		},
+	}
+	workConsumed := factorytoken.Token{
+		ID:        "tok-1",
+		PlaceID:   "wt-code:init",
+		CreatedAt: now.Add(-time.Hour),
+		EnteredAt: now.Add(-time.Hour),
+		Color: factorytoken.Color{
+			WorkID:     "w-resource-failure",
+			WorkTypeID: "wt-code",
+		},
+		History: factorytoken.History{
+			TotalVisits:         map[string]int{},
+			ConsecutiveFailures: map[string]int{},
+			PlaceVisits:         map[string]int{},
+		},
+	}
+
+	consumedTokens := []factorytoken.Token{resourceConsumed, workConsumed}
+	if workFirst {
+		consumedTokens = []factorytoken.Token{workConsumed, resourceConsumed}
+	}
+
+	snapshot := &interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net]{
+		Marking: petri.MarkingSnapshot{
+			Tokens: map[string]*factorytoken.Token{
+				workConsumed.ID:     &workConsumed,
+				resourceConsumed.ID: &resourceConsumed,
+			},
+			PlaceTokens: map[string][]string{
+				"wt-code:init":       {workConsumed.ID},
+				"executor:available": {resourceConsumed.ID},
+			},
+		},
+		Dispatches: map[string]*interfaces.DispatchEntry{
+			"d-1": {
+				DispatchID:     "d-1",
+				TransitionID:   "t1",
+				ConsumedTokens: factorytoken.ToWorkerSlice(consumedTokens),
+			},
+		},
+	}
+	return tp, snapshot, resourceConsumed
+}
+
+func assertFailedMixedWorkResourceRelease(t *testing.T, result *interfaces.TickResult, resourceConsumed factorytoken.Token) {
+	t.Helper()
+	if result == nil || len(result.Mutations) != 2 {
+		t.Fatalf("expected 2 mutations, got %+v", result)
+	}
+
+	var failedWork *interfaces.MarkingMutation
+	var releasedResource *interfaces.MarkingMutation
+	for i := range result.Mutations {
+		if result.Mutations[i].NewToken == nil {
+			continue
+		}
+		switch result.Mutations[i].NewToken.Color.DataType {
+		case factorytoken.DataTypeWork:
+			failedWork = &result.Mutations[i]
+		case factorytoken.DataTypeResource:
+			releasedResource = &result.Mutations[i]
+		}
+	}
+	if failedWork == nil {
+		t.Fatal("expected failed work mutation")
+	}
+	if failedWork.ToPlace != "wt-code:failed" {
+		t.Fatalf("failed work ToPlace = %q, want wt-code:failed", failedWork.ToPlace)
+	}
+	if failedWork.NewToken.Color.WorkID != "w-resource-failure" {
+		t.Fatalf("failed work WorkID = %q, want w-resource-failure", failedWork.NewToken.Color.WorkID)
+	}
+	if failedWork.NewToken.History.LastError != "agent crashed" {
+		t.Fatalf("failed work LastError = %q, want agent crashed", failedWork.NewToken.History.LastError)
+	}
+	assertReleasedResourceMutation(t, releasedResource, resourceConsumed)
+}
+
+func assertReleasedResourceMutation(t *testing.T, released *interfaces.MarkingMutation, resourceConsumed factorytoken.Token) {
+	t.Helper()
+	if released == nil {
+		t.Fatal("expected released resource mutation")
+	}
+	if released.ToPlace != "executor:available" {
+		t.Fatalf("ToPlace = %q, want %q", released.ToPlace, "executor:available")
+	}
+	if released.NewToken.ID != resourceConsumed.ID || released.NewToken.Color.WorkID != resourceConsumed.Color.WorkID {
+		t.Fatalf("released resource token = %#v, want preserved identity from %#v", released.NewToken, resourceConsumed)
+	}
+	if !released.NewToken.CreatedAt.Equal(resourceConsumed.CreatedAt) {
+		t.Fatalf("CreatedAt = %v, want %v", released.NewToken.CreatedAt, resourceConsumed.CreatedAt)
+	}
+	if released.NewToken.Color.Tags["pool"] != "shared" {
+		t.Fatalf("tag pool = %q, want %q", released.NewToken.Color.Tags["pool"], "shared")
+	}
+	if released.NewToken.History.PlaceVisits["executor:available"] != 4 {
+		t.Fatalf("PlaceVisits = %#v, want preserved history", released.NewToken.History.PlaceVisits)
+	}
+}
+
+func TestTransitioner_LateDispatchResultForTerminalOrFailedWorkIsIgnored(t *testing.T) {
+	now := time.Date(2026, time.August, 29, 12, 30, 0, 0, time.UTC)
+	for _, test := range []struct {
+		name      string
+		placeID   string
+		stateName string
+		stateType interfaces.StateType
+	}{
+		{name: "terminal", placeID: "wt-code:done", stateName: "done", stateType: interfaces.StateTypeTerminal},
+		{name: "failed", placeID: "wt-code:failed", stateName: "failed", stateType: interfaces.StateTypeFailed},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			assertIgnoredTerminalResult(t, now, test.placeID, test.name, test.stateName, test.stateType)
+		})
+	}
+}
+
+func assertIgnoredTerminalResult(t *testing.T, now time.Time, placeID, resultName, stateName string, stateType interfaces.StateType) {
+	t.Helper()
+	net := buildPipelineNet()
+	transitioner := NewTransitioner(net, nil, func() time.Time { return now }, testTokenTransformer(net), nil, nil, nil, testWorkPropagationPolicy())
+	snapshot := pipelineSnapshot(
+		placeID,
+		"t1",
+		"d-late-"+resultName,
+		factorytoken.Color{WorkID: "w-terminal", WorkTypeID: "wt-code"},
+		now,
+	)
+	snapshot.Results = []workerexecution.WorkResult{{
+		DispatchID: "d-late-" + resultName, TransitionID: "t1", Outcome: workerexecution.OutcomeAccepted,
+		Output: "worker output must not be retained", Error: "worker error must not be retained",
+	}}
+
+	result, err := transitioner.Execute(context.Background(), &snapshot)
+	if err != nil {
+		t.Fatalf("Execute() error = %v", err)
+	}
+	if result == nil {
+		t.Fatal("Execute() result = nil, want retired ignored dispatch")
+	}
+	if len(result.Mutations) != 0 || len(result.GeneratedBatches) != 0 {
+		t.Fatalf("ignored result changed runtime = mutations:%#v generated:%#v", result.Mutations, result.GeneratedBatches)
+	}
+	if len(result.CompletedDispatches) != 1 {
+		t.Fatalf("completed dispatches = %#v, want one ignored dispatch", result.CompletedDispatches)
+	}
+	completed := result.CompletedDispatches[0]
+	if completed.IgnoredWorkID != "w-terminal" || completed.IgnoredResult == nil {
+		t.Fatalf("ignored completion = %#v, want Work w-terminal marker", completed)
+	}
+	ignored := completed.IgnoredResult
+	if ignored.Reason != interfaces.DispatchResultIgnoredReasonWorkAlreadyTerminal ||
+		ignored.ResultOutcome != workerexecution.OutcomeAccepted ||
+		ignored.ObservedState.Name != stateName || ignored.ObservedState.Type != stateType {
+		t.Fatalf("ignored payload = %#v, want %s/%s accepted", ignored, stateName, stateType)
+	}
+}
+
+func TestTransitioner_LateDispatchResultGuardsMultiInputResultAtomically(t *testing.T) {
+	now := time.Date(2026, time.August, 29, 12, 31, 0, 0, time.UTC)
+	net := buildPipelineNet()
+	transitioner := NewTransitioner(net, nil, func() time.Time { return now }, testTokenTransformer(net), nil, nil, nil, testWorkPropagationPolicy())
+	snapshot := pipelineSnapshot(
+		"wt-code:done", "t1", "d-multi-late",
+		factorytoken.Color{WorkID: "w-terminal", WorkTypeID: "wt-code"}, now,
+	)
+	second := factorytoken.Token{
+		ID: "tok-2", PlaceID: "wt-code:init", CreatedAt: now, EnteredAt: now,
+		Color: factorytoken.Color{WorkID: "w-processing", WorkTypeID: "wt-code"},
+	}
+	snapshot.Marking.Tokens[second.ID] = &second
+	snapshot.Marking.PlaceTokens[second.PlaceID] = []string{second.ID}
+	first := *snapshot.Marking.Tokens["tok-1"]
+	snapshot.Dispatches["d-multi-late"].ConsumedTokens = factorytoken.ToWorkerSlice([]factorytoken.Token{first, second})
+	snapshot.Results = []workerexecution.WorkResult{{
+		DispatchID: "d-multi-late", TransitionID: "t1", Outcome: workerexecution.OutcomeFailed,
+		Output: "must not generate or route", Error: "late failure",
+	}}
+
+	result, err := transitioner.Execute(context.Background(), &snapshot)
+	if err != nil {
+		t.Fatalf("Execute() error = %v", err)
+	}
+	if result == nil || len(result.Mutations) != 0 || len(result.GeneratedBatches) != 0 || len(result.CompletedDispatches) != 1 {
+		t.Fatalf("atomic ignored result = %#v, want no mutations/generated and one completion", result)
+	}
+	completed := result.CompletedDispatches[0]
+	if completed.IgnoredWorkID != "w-terminal" || completed.IgnoredResult == nil ||
+		completed.IgnoredResult.ObservedState.Type != interfaces.StateTypeTerminal {
+		t.Fatalf("multi-input ignored completion = %#v, want first terminal Work marker", completed)
+	}
+}
