+++ pkg/transports/http/contracttests/openapi_contract_events_test.go
@@ -22,6 +22,45 @@ var retiredFactoryEventTypeValues = []string{
 	"RUN_FINISHED",
 }
 
+var canonicalFactoryEventTypeValues = []string{
+	"RUN_REQUEST",
+	"INITIAL_STRUCTURE_REQUEST",
+	"FACTORY_CHANGE",
+	"FACTORY_CHANGE_REQUEST",
+	"FACTORY_CHANGE_FAILED",
+	"WORK_REQUEST",
+	"RELATIONSHIP_CHANGE_REQUEST",
+	"DISPATCH_REQUEST",
+	"HUMAN_APPROVAL_REQUESTED",
+	"DISPATCH_WORKER_SESSION_ASSOCIATION",
+	"MODEL_REQUEST",
+	"MODEL_RESPONSE",
+	"INFERENCE_REQUEST",
+	"INFERENCE_RESPONSE",
+	"SCRIPT_REQUEST",
+	"SCRIPT_RESPONSE",
+	"AGENT_RUN_RESPONSE",
+	"DISPATCH_RESPONSE",
+	"DISPATCH_RESULT_IGNORED",
+	"WORK_STATE_CHANGE",
+	"FACTORY_STATE_RESPONSE",
+	"RUN_RESPONSE",
+	"SESSION_STARTED",
+	"SESSION_PAUSED",
+	"SESSION_RESUMED",
+	"SESSION_RESULT_UPDATED",
+	"SESSION_COMPLETED",
+	"SESSION_LIFECYCLE_CONTROL",
+	"ORCHESTRATOR_PHASE_CHANGED",
+	"ORCHESTRATOR_CHECKPOINT_WRITTEN",
+	"DISPATCH_QUEUED",
+	"DISPATCH_INTERRUPTED",
+	"DISPATCH_RECONCILED",
+	"JAVASCRIPT_CHECKPOINT_REF",
+	"JAVASCRIPT_PHASE_CHANGE",
+	"ARTIFACT_CREATED",
+}
+
 func TestOpenAPIContract_DefinesUnifiedFactoryEventLog(t *testing.T) {
 	doc := loadBundledOpenAPIDocument(t)
 	schemas := componentSchemas(t, doc)
